requirements.txt
aws-cdk-lib==2.148.0
constructs>=10.0.0,<11.0.0
streamlit
Pillow
boto3>=1.28.57
awscli>=1.29.57
botocore>=1.31.57
langchain
faiss-cpu>=1.7,<2
pypdf>=3.8,<4
sqlalchemy
cdk-nag